fix(gateway): reconcile pairing lifecycle after token removal
