+++ src/gateway/server-methods/nodes.pairing.ts
@@ -139,19 +139,11 @@ function emitNodeRoleRemovalSecurityEvent(params: {
 async function removePairedDeviceBackedNode(params: {
   nodeId: string;
   client: GatewayClient | null;
-  context: Pick<
-    GatewayRequestContext,
-    | "disconnectClientsForDevice"
-    | "invalidateClientsForDevice"
-    | "logGateway"
-    | "workerEnvironmentService"
-    | "workerPlacementDispatchService"
-  >;
+  context: Pick<GatewayRequestContext, "invalidateClientsForDevice" | "logGateway">;
 }): Promise<
   | {
       status: "removed";
       nodeId: string;
-      disconnectDeviceId: string;
     }
   | { status: "denied"; message: string }
   | { status: "unknown" }
@@ -210,11 +202,9 @@ async function removePairedDeviceBackedNode(params: {
     role: "node",
     reason: "device-pair-removed",
   });
-  await reconcileRevokedDeviceWorker(params.context, removed.deviceId);
   return {
     status: "removed",
     nodeId: removed.deviceId,
-    disconnectDeviceId: removed.deviceId,
   };
 }
 
@@ -376,7 +366,7 @@ export const nodePairingHandlers: GatewayRequestHandlers = {
     });
   },
   // Remove a node pairing (CLI: `openclaw nodes remove`). This revokes the
-  // device's `node` role in devices/paired.json, which drops the approved node
+  // device's `node` role in the paired-device store, which drops the approved node
   // surface with it, and disconnects the device's node-role sessions: a
   // mixed-role device keeps its row and only loses the `node` role, a
   // node-only device row is deleted. Authz mirrors device.pair.remove:
@@ -400,13 +390,14 @@ export const nodePairingHandlers: GatewayRequestHandlers = {
       }
       try {
         clearRemovedNodeRuntimeState({ nodeId: deviceBacked.nodeId, context });
+        await reconcileRevokedDeviceWorker(context, deviceBacked.nodeId);
         broadcastRemovedNodePairing({ nodeId: deviceBacked.nodeId, context });
         respond(true, { nodeId: deviceBacked.nodeId }, undefined);
       } finally {
         // Preserve response-first shutdown on success, while guaranteeing the
         // hard close when runtime cleanup or later bookkeeping throws.
         queueMicrotask(() => {
-          context.disconnectClientsForDevice?.(deviceBacked.disconnectDeviceId, {
+          context.disconnectClientsForDevice?.(deviceBacked.nodeId, {
             role: "node",
           });
         });
